# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for tracking daily activities, streaks, and analytics.\n\n## Features\n- User authentication with JWT tokens\n- Activity tracking with customizable types\n- Streak tracking to maintain consistency\n- Weekly analytics with daily breakdowns\n- Privacy controls for user profiles\n- Tile configuration for dashboard customization\n\n## Rate Limits\n- Login/Register: 5 requests per minute per IP\n- Password Reset: 3 requests per minute per IP\n- General API: 60 requests per minute per user",
        "title": "Growth Tracker API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/aman1117/growth-tracker"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/auth/forgot-password": {
            "post": {
                "description": "Send password reset email (always returns success for security)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Request password reset",
                "parameters": [
                    {
                        "description": "User email",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ForgotPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reset email sent (if user exists)",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/auth/reset-password": {
            "post": {
                "description": "Reset password using the token from email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Reset password with token",
                "parameters": [
                    {
                        "description": "Reset token and new password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password reset successful",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid token or validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/change-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Change the authenticated user's password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Change password",
                "parameters": [
                    {
                        "description": "Current and new password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password changed successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error or incorrect password",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "description": "Authenticate user with email/username and password, returns JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful with JWT token",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "description": "Create a new user account with email, username, and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Registration details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error or user exists",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/update-username": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the authenticated user's username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Update username",
                "parameters": [
                    {
                        "description": "New username",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.UpdateUsernameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Username updated successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.UsernameUpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error or username taken",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_aman1117_backend_internal_dto.ChangePasswordRequest": {
            "description": "Change password for authenticated user",
            "type": "object",
            "properties": {
                "current_password": {
                    "type": "string",
                    "example": "OldPass123"
                },
                "new_password": {
                    "type": "string",
                    "example": "NewPass123"
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.ErrorResponse": {
            "description": "Generic error response",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "error_code": {
                    "type": "string",
                    "example": "INVALID_REQUEST"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "github_com_aman1117_backend_internal_dto.ForgotPasswordRequest": {
            "description": "Password reset request initiation",
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.LoginRequest": {
            "description": "User login request",
            "type": "object",
            "properties": {
                "identifier": {
                    "description": "Can be email or username",
                    "type": "string",
                    "example": "john_doe"
                },
                "password": {
                    "type": "string",
                    "example": "SecurePass123"
                }
            }
        },
        "github_com_aman1117_backend_internal_dto.LoginResponse": {
            "description": "Successful login response with JWT token",
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIs..."
                },
                "expires_at": {
                    "type": "string",
                    "example": "2026-01-05T12:00:00Z"
                },
                "expires_in": {
                    "description": "seconds",
                    "type": "integer",
                    "example": 86400
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.RegisterRequest": {
            "description": "User registration request",
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "SecurePass123"
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                }
            }
        },
        "github_com_aman1117_backend_internal_dto.ResetPasswordRequest": {
            "description": "Password reset with token",
            "type": "object",
            "properties": {
                "confirm_password": {
                    "type": "string",
                    "example": "NewSecurePass123"
                },
                "new_password": {
                    "type": "string",
                    "example": "NewSecurePass123"
                },
                "token": {
                    "type": "string",
                    "example": "abc123def456"
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.SuccessResponse": {
            "description": "Generic success response",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.UpdateUsernameRequest": {
            "description": "Username update request",
            "type": "object",
            "properties": {
                "new_username": {
                    "type": "string",
                    "example": "new_username"
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.UsernameUpdateResponse": {
            "description": "Username update result",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Username updated successfully"
                },
                "new_username": {
                    "type": "string",
                    "example": "new_username"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}